Reduce memory usage on ingest
This commit reduces memory usage by ensuring WriteRequests and Labels
are freed as soon as they are no longer needed. For WriteRequests, this
is immediately after parseData is called, as at that point we are only
going to used the parsed version. For Labels this is as soon as we have
the SeriesId for the Labels.

This commit also adds a command line flag for async-acks, and the
ability ro report the actual throughput of said acks.