split pool part 1: WriteRequests
The two things stored in an InsertContext have vastly different
lifetimes: the WriteRequest is unneeded as soon as ParseData is done,
while the Labels should be freed at some point during the actual series-
insertion. This commit is the first part of splitting the InstertCtx
pool in two. It creates and uses a dedicated WriteRequest pool. While
the Label pool is also created, it is not yet used, as we need to change
to storing *Label instead of unboxed Label everywhere to really see a
benefit.